// src/enums.ts
export enum Symbols {
  AAPL = "AAPL",
  MSFT = "MSFT",
  LMT = "LMT",
  TSLA = "TSLA",
  NVDA = "NVDA",
  DPZ = "DPZ",
  CAT = "CAT",
  EA = "EA",
  NFLX = "NFLX",
  SPCE = "SPCE",
}

export enum Products {
  Apple = "iPhone",
  Microsoft = "MS Office 365",
  Lewis = "LMT MARS",
  Tesla = "Tesla Model S",
  Nvidia = "Graphics cards",
  Dominos = "Delivery by robots",
  Caterpillar = "Crawler excavator",
  ElectronicArts = "FIFA game series",
  Netflix = "Movies and series",
  Virgin = "Private spacecraft",
}

export enum FullNameOfCompanies {
  Apple = "Apple Inc.",
  Microsoft = "Microsoft Corporation",
  Lewis = "Lockheed Martin Corporation",
  Tesla = "Tesla, Inc.",
  Nvidia = "Nvidia Corporation",
  Dominos = "Domino's Pizza, Inc.",
  Caterpillar = "Caterpillar Inc.",
  ElectronicArts = "Electronic Arts",
  Netflix = "Netflix, Inc.",
  Virgin = "Virgin Galactic",
}